名前を保存して各箇所で表示 refs #54
