+++ public/local-storage.html
@@ -29,14 +29,25 @@
     <!-- /.header -->
     <main class="main">
       <div class="container">
-        key:
-        <input type="text" id="key"> value:
-        <input type="text" id="value">
-        <input type="submit" id="put" value="Put">
-        <input type="submit" id="remove" value="Remove">
-        <input type="submit" id="clear" value="Clear All">
-        <div id="result"></div>
-        <!-- /#result -->
+        <div class="row">
+          <div class="col-xs-12">
+            value:
+            <input type="text" id="value">
+            <input type="submit" id="save" value="Save">
+            <input type="submit" id="remove" value="Remove">
+          </div>
+          <!-- /.col-xs-12 -->
+        </div>
+        <!-- /.row -->
+        <div class="row">
+          <div class="col-xs-12">
+            こんにちは
+            <span class="result"></span>さん。はじめまして。
+            <span class="result"></span>さん。ごきげんいかがですか。
+          </div>
+          <!-- /.col-xs-12 -->
+        </div>
+        <!-- /.row -->
       </div>
       <!-- /.container -->
     </main>
